Add relay to system
The main program does now have a button for turning on and a button for turning off.
When the program is turned on, a led lights up and the relay is set to high which in turn opens the solenoide.
Currently, the interval is hardcoded with 1 sec on and two sec pause.
By clicking the turn off button, the whole program is turned off. By clicking the turn on button, the loop will start over.

Sketch of total system is added.